User Story:1185 - API Development for Reject Finance
git-svn-id: http://192.168.110.32/svn/Products/PFF/core/branches/mlod@103351 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/finance/impl/FinanceDetailServiceImpl.java
@@ -5738,7 +5738,7 @@ public class FinanceDetailServiceImpl extends GenericFinanceDetailService implem
 
 		// Save Finance Details Data on Reject Tables
 		// =======================================
-		if (!isWIF && StringUtils.equals(financeMain.getRecordType(), PennantConstants.RECORD_TYPE_NEW)) {
+		if (!isWIF && StringUtils.equals(financeMain.getRecordType(), PennantConstants.RECORD_TYPE_NEW)) {		
 			getFinanceMainDAO().saveRejectFinanceDetails(financeMain);
 		}
 
@@ -5844,7 +5844,10 @@ public class FinanceDetailServiceImpl extends GenericFinanceDetailService implem
 
 		// Finance Main Details Deletion
 		// =======================================
-		getFinanceMainDAO().delete(financeMain, TableType.TEMP_TAB, isWIF, isAutoReject);
+		FinanceMain dbFinanceMain = getFinanceMain(financeDetail.getFinScheduleData().getFinReference(), "_Temp");
+		if (!(null == dbFinanceMain)) {
+			getFinanceMainDAO().delete(financeMain, TableType.TEMP_TAB, isWIF, isAutoReject);
+		}
 
 		auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
 		String[] fields = PennantJavaUtil.getFieldDetails(new FinanceMain(), financeMain.getExcludeFields());
